Fix bugs in static constructor, adjust tests
In .NET Core 2.1 and above, the signature of the TypeArguments property of the internal ICSharpInvokeOrInvokeMemberBinder interface changed from IList<Type> to Type[]. What this boils down to is that generic arguments were not being captured in a small number of cases.

In the .NET Core 3.0, static readonly fields are enforced by reflection. This caused the static constructor of UniversalMemberAccessor to throw a TypeInitializationException when it performed feature detection on whether modifying readonly fields is supported. The solution is to wrap the feature detection in try/catch blocks.